+++ tests/resources/test_project/src/moduleA/submoduleA1/submoduleA11/fileA11.py
@@ -1,4 +1,5 @@
 from src.moduleB.submoduleB1.fileB1 import b1  # type: ignore
+import os  # noqa: F401 E401
 
 
 def a11():